backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
websockets==14.1
google-generativeai==0.8.3
google-genai>=0.6.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20
python-dotenv==1.0.1
aiofiles==24.1.0
